requirements.txt
flask
three
pytesseract
pillow